Move TagsGetter to oci package
