+++ internal/oci/mock.go
@@ -3,6 +3,7 @@ package oci
 import (
 	"context"
 
+	"github.com/artifacthub/hub/internal/hub"
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
 	"github.com/stretchr/testify/mock"
 )
@@ -24,4 +25,16 @@ func (m *PullerMock) PullLayer(
 	desc, _ := args.Get(0).(ocispec.Descriptor)
 	data, _ := args.Get(1).([]byte)
 	return desc, data, args.Error(2)
+}
+
+// TagsGetterMock is a mock implementation of the hub.OCITagsGetter interface.
+type TagsGetterMock struct {
+	mock.Mock
+}
+
+// Tags implements the OCITagsGetter interface.
+func (m *TagsGetterMock) Tags(ctx context.Context, r *hub.Repository) ([]string, error) {
+	args := m.Called(ctx, r)
+	tags, _ := args.Get(0).([]string)
+	return tags, args.Error(1)
 }
